Modificaciones en MascotaImpl - Borrado de HigieneImpl
Se han realizado las siguientes modificaciones:

* Se han borrado las funciones save() y reload() al carecer de uso en la clase MascotaImpl.

* Se ha borrado la clase HigieneImpl, se ha agregado directamente a MascotaImpl implementando tanto Higiene como Mascota. 

* Se ha arreglado el problema con la variable grupo_sanguineo, modificandolo para establecer el uso del camelCase (grupoSanguineo)

* Se ha borrado la función getInstance ya que carece de sentido en MascotasImpl.

* Se ha modificado la función isValidGrupoSanguineo, borrando el caso de hacer minusculas las palabras que proporcione el usuario.
