+++ src/main/java/org/balanpy/Higiene.java
@@ -1,12 +1,8 @@
 package org.balanpy;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Map;
-
-import com.fasterxml.jackson.core.exc.StreamReadException;
-import com.fasterxml.jackson.databind.DatabindException;
 
 public interface Higiene {
 
@@ -33,10 +29,4 @@ public interface Higiene {
 	public ArrayList<String> getProductosNR(); // GETTER
 
 	public void setProductosNR(ArrayList<String> productosNR); // SETTER
-
-	// Helper function for persisting user information
-	public void saveM() throws StreamReadException, DatabindException, IOException;
-
-	// Helper for reloading the object
-	public void reloadM() throws StreamReadException, DatabindException, IOException;
 }
